requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.8.1
attrs==23.2.0
binutils==0.1.dev0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.1.0
Django==4.2.11
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
django-rest-knox @ git+https://github.com/jazzband/django-rest-knox.git@09403255a069c2416a5dfb889f80342a2d7b2646
django-twilio==0.14.3.3
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.4.0
filelock==3.13.3
frozenlist==1.4.1
glibc==0.6.1
idna==3.4
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
oauthlib==3.2.2
phonenumbers==8.13.34
phonenumberslite==8.13.34
pillow==10.3.0
platformdirs==4.2.0
pycparser==2.22
pyinstaller==4.0
pyinstaller-hooks-contrib==2022.12
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pip==1.1.1
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==2.0.0
semantic-version==2.10.0
setuptools-rust==1.9.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
twilio==9.0.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.12
virtualenv==20.25.1
yarl==1.9.4
